Add code to interpolate lat-lon to se grid for DeepONet
This commits add a subroutine in  nudging.F90 to initilize the weighting
for linear interpolation from lat-lon to model grid. The idea is to
initialize the weights once, and applied it for model calculation so
that the model will not need to calculate the weights at each time step.
The goal is to reduce the computational time