Code Changes to implement online DeepONet machine learning bias correction
This commit contains the code changes in physpkg.F90 and nudging.F90
that are used to implement the online DeepONet machine learning(ML)
bias correction model. Two major changes are made:
1. The changes in physpkg.F90 is made to let E3SM call the DeepONet
   model using a)model state after dynamics and b) model state after
   the radiation parameterizations. These two options exist in previous
   nudging implementation to test the sensitivity of nudging to input
   model state provided at different locations
2. The changes in nudging.F90 is made to add subroutines to a)
   interpolate model state from SE grid to lat-lon grid at limited area
   which is used by the convolution 2D DeepONet model and b) call
   DeepONet machine learning model to predict the nudging tendency
   and then return it back to model state

The changes in namelist_definition.xml and config_machines.xml are also
made to include the namelist variables to controling the way of runing
DeepONet machine learning model, and the libraries required for
compiling the pytorch and pytorch-c-fortran bidings.